+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add --no-config to ripgrep command to prevent parsing of user rg.conf files
 - Fix object migration bug
 - Improve folder globs for windows
+- Replace all instances of tags in regex (thanks to [tylerbrockett](https://github.com/tylerbrockett))
 
 # v0.0.153 - 2019-09-12
 - Update README.md with new settings (thanks to [abelmatser](https://github.com/abelmatser))